sign and broadcast the vote tx
